+++ src/html/html-script-tag.ts
@@ -34,7 +34,7 @@ export class ScriptTagBackReferenceImport extends Import {
 export class ScannedScriptTagImport extends ScannedImport {
   resolve(document: Document): ScriptTagImport|undefined {
     const resolvedUrl = document._analysisContext.resolver.resolve(
-        this.url, document.parsedDocument.baseUrl, this);
+        document.parsedDocument.baseUrl, this.url, this);
     if (resolvedUrl === undefined) {
       return;
     }